requirements.txt
torch
ollama
chromadb
langchain
sentence-transformers
transformers
PyPDF2
python-docx
numpy
scikit-learn
pandas